requirements.txt
Flask~=1.1.2
rasa[spacy]~=2.5.1
requests~=2.25.1
flask-cors~=3.0.10
python-dateutil~=2.8.1
pymorphy2~=0.9.1
pymorphy2-dicts-uk~=2.4.1